requirements.txt
Django==5.0
psycopg2-binary==2.9.6
djangorestframework==3.14
pytest-django